+++ test/test_helper.exs
@@ -6,6 +6,7 @@ ExUnit.start(
     :integration_bolt_tls,
     :integration_shutdown,
     :integration_ts,
-    :integration_grpc
+    :integration_grpc,
+    :integration_grpc_tls
   ]
 )